Raising Value and Index Errors due to inappropriate data structure
This is an attempt to recreate the original version of incremental beam search. It is raising errors because in order to maintain monotonicity, duplicates must be kept in the waitlist, and the heap structure used as the foundation of the waitlist handles duplicates through overwriting them. This algorithm should be workable with a red-black tree or similar structure used for the waitlist and the current paired heaps used for the openlist.