Non-monotonic variant of IBS
Due to its checking the waitlist for duplicates, this algorithm is fundamentally non-monotonic. Unfortunately, its data structures are not designed to hold duplicates in the waitlist, as is necessary to maintain monotonicity. However, its goal pruning still represents a feasible strategy, provided the waitlist was altered to use a red-black tree or similar structure. It also now transfers the best item from the waitlist to the openlist, rather than the worst.